+++ CHANGES.md
@@ -1,3 +1,6 @@
+Changes in version 2021-06-20
+* Fix Snakes: Shuffles answer in multichoice questions
+
 Changes in version 2021-05-31
 * Fix hiddenpicture: Problem when user gives wrong answer